Restructured files for CLI
