+++ cli/cli.class.php
@@ -1,8 +1,11 @@
 <?php
 if ( ! defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
-require_once 'cli-extras.class.php';
-require_once 'cli-commands.class.php';
+require_once __DIR__ . '/cli-extras.class.php';
+require_once __DIR__ . '/classes/general.php';
+require_once __DIR__ . '/classes/cloudflare.php';
+require_once __DIR__ . '/classes/optimizations.php';
+require_once __DIR__ . '/classes/fpc.php';
 
 /**
  * Class Servebolt_CLI
@@ -10,7 +13,7 @@ require_once 'cli-commands.class.php';
  *
  * Does all the WP CLI handling.
  */
-class Servebolt_CLI extends Servebolt_CLI_Commands {
+class Servebolt_CLI {
 
 	/**
 	 * Singleton instance.
@@ -43,57 +46,57 @@ class Servebolt_CLI extends Servebolt_CLI_Commands {
 	 */
 	private function registerCommands() {
 
+		$general        = new Servebolt_CLI_General;
+		$optimizations  = new Servebolt_CLI_Optimizations;
+		$fpc            = new Servebolt_CLI_FPC;
+		$cf             = new Servebolt_CLI_Cloudflare;
+
 		// General
-		WP_CLI::add_command( 'servebolt delete-all-settings',         [$this, 'command_delete_all_settings'] );
+		WP_CLI::add_command( 'servebolt delete-all-settings',        [$general, 'command_delete_all_settings'] );
 
 		// Optimization
-		WP_CLI::add_command( 'servebolt db optimize',                [$this, 'command_optimize_database'] );
-		WP_CLI::add_command( 'servebolt db fix',                     [$this, 'command_fix'] );
-		WP_CLI::add_command( 'servebolt db analyze',                 [$this, 'command_analyze_tables'] );
+		WP_CLI::add_command( 'servebolt db optimize',                [$optimizations, 'command_optimize_database'] );
+		WP_CLI::add_command( 'servebolt db fix',                     [$optimizations, 'command_fix'] );
+		WP_CLI::add_command( 'servebolt db analyze',                 [$optimizations, 'command_analyze_tables'] );
 
 		// Servebolt Full Page Cache
-		WP_CLI::add_command( 'servebolt fpc activate',               [$this, 'command_nginx_fpc_enable'] );
-		WP_CLI::add_command( 'servebolt fpc deactivate',             [$this, 'command_nginx_fpc_disable'] );
-		WP_CLI::add_command( 'servebolt fpc status',                 [$this, 'command_nginx_fpc_status'] );
+		WP_CLI::add_command( 'servebolt fpc activate',               [$fpc, 'command_nginx_fpc_enable'] );
+		WP_CLI::add_command( 'servebolt fpc deactivate',             [$fpc, 'command_nginx_fpc_disable'] );
+		WP_CLI::add_command( 'servebolt fpc status',                 [$fpc, 'command_nginx_fpc_status'] );
 
-		WP_CLI::add_command( 'servebolt fpc post-types get',         [$this, 'command_nginx_fpc_get_cache_post_types'] );
-		WP_CLI::add_command( 'servebolt fpc post-types set',         [$this, 'command_nginx_fpc_set_cache_post_types'] );
-		WP_CLI::add_command( 'servebolt fpc post-types clear',       [$this, 'command_nginx_fpc_clear_cache_post_types'] );
+		WP_CLI::add_command( 'servebolt fpc post-types get',         [$fpc, 'command_nginx_fpc_get_cache_post_types'] );
+		WP_CLI::add_command( 'servebolt fpc post-types set',         [$fpc, 'command_nginx_fpc_set_cache_post_types'] );
+		WP_CLI::add_command( 'servebolt fpc post-types clear',       [$fpc, 'command_nginx_fpc_clear_cache_post_types'] );
 
-		WP_CLI::add_command( 'servebolt fpc excluded-posts get',     [$this, 'command_nginx_fpc_get_excluded_posts'] );
-		WP_CLI::add_command( 'servebolt fpc excluded-posts set',     [$this, 'command_nginx_fpc_set_excluded_posts'] );
-		WP_CLI::add_command( 'servebolt fpc excluded-posts clear',   [$this, 'command_nginx_fpc_clear_excluded_posts'] );
+		WP_CLI::add_command( 'servebolt fpc excluded-posts get',     [$fpc, 'command_nginx_fpc_get_excluded_posts'] );
+		WP_CLI::add_command( 'servebolt fpc excluded-posts set',     [$fpc, 'command_nginx_fpc_set_excluded_posts'] );
+		WP_CLI::add_command( 'servebolt fpc excluded-posts clear',   [$fpc, 'command_nginx_fpc_clear_excluded_posts'] );
 
 		// Cloudflare
-		//WP_CLI::add_command( 'servebolt cf setup',                   [$this, 'command_cf_setup'] ); // TODO: Make interactive setup command
-		WP_CLI::add_command( 'servebolt cf status',                  [$this, 'command_cf_status'] );
-		WP_CLI::add_command( 'servebolt cf activate',                [$this, 'command_cf_enable'] );
-		WP_CLI::add_command( 'servebolt cf deactivate',              [$this, 'command_cf_disable'] );
-		WP_CLI::add_command( 'servebolt cf get-config',              [$this, 'command_cf_get_config'] );
+		//WP_CLI::add_command( 'servebolt cf setup',                   [$cf, 'command_cf_setup'] ); // TODO: Make interactive setup command
+		WP_CLI::add_command( 'servebolt cf status',                  [$cf, 'command_cf_status'] );
+		WP_CLI::add_command( 'servebolt cf activate',                [$cf, 'command_cf_enable'] );
+		WP_CLI::add_command( 'servebolt cf deactivate',              [$cf, 'command_cf_disable'] );
+		WP_CLI::add_command( 'servebolt cf get-config',              [$cf, 'command_cf_get_config'] );
 
-		WP_CLI::add_command( 'servebolt cf api test',                [$this, 'command_cf_test_api_connection'] );
-		WP_CLI::add_command( 'servebolt cf api credentials get',     [$this, 'command_cf_get_credentials'] );
-		WP_CLI::add_command( 'servebolt cf api credentials set',     [$this, 'command_cf_set_credentials'] );
-		WP_CLI::add_command( 'servebolt cf api credentials clear',   [$this, 'command_cf_clear_credentials'] );
+		WP_CLI::add_command( 'servebolt cf api test',                [$cf, 'command_cf_test_api_connection'] );
+		WP_CLI::add_command( 'servebolt cf api credentials get',     [$cf, 'command_cf_get_credentials'] );
+		WP_CLI::add_command( 'servebolt cf api credentials set',     [$cf, 'command_cf_set_credentials'] );
+		WP_CLI::add_command( 'servebolt cf api credentials clear',   [$cf, 'command_cf_clear_credentials'] );
 
-		WP_CLI::add_command( 'servebolt cf zone list',               [$this, 'command_cf_list_zones'] );
-		WP_CLI::add_command( 'servebolt cf zone get',                [$this, 'command_cf_get_zone'] );
-		WP_CLI::add_command( 'servebolt cf zone set',                [$this, 'command_cf_set_zone'] );
-		WP_CLI::add_command( 'servebolt cf zone clear',              [$this, 'command_cf_clear_zone'] );
+		WP_CLI::add_command( 'servebolt cf zone list',               [$cf, 'command_cf_list_zones'] );
+		WP_CLI::add_command( 'servebolt cf zone get',                [$cf, 'command_cf_get_zone'] );
+		WP_CLI::add_command( 'servebolt cf zone set',                [$cf, 'command_cf_set_zone'] );
+		WP_CLI::add_command( 'servebolt cf zone clear',              [$cf, 'command_cf_clear_zone'] );
 
-		WP_CLI::add_command( 'servebolt cf purge type',              [$this, 'command_cf_set_purge_type'] );
-		WP_CLI::add_command( 'servebolt cf purge status',            [$this, 'command_cf_purge_status'] );
+		WP_CLI::add_command( 'servebolt cf purge type',              [$cf, 'command_cf_set_purge_type'] );
+		WP_CLI::add_command( 'servebolt cf purge status',            [$cf, 'command_cf_purge_status'] );
 
-		WP_CLI::add_command( 'servebolt cf purge clear-queue',       [$this, 'command_cf_clear_cache_purge_queue'] );
-		WP_CLI::add_command( 'servebolt cf purge url',               [$this, 'command_cf_purge_url'] );
-		WP_CLI::add_command( 'servebolt cf purge post',              [$this, 'command_cf_purge_post'] );
-		WP_CLI::add_command( 'servebolt cf purge all',               [$this, 'command_cf_purge_all'] );
+		WP_CLI::add_command( 'servebolt cf purge clear-queue',       [$cf, 'command_cf_clear_cache_purge_queue'] );
+		WP_CLI::add_command( 'servebolt cf purge url',               [$cf, 'command_cf_purge_url'] );
+		WP_CLI::add_command( 'servebolt cf purge post',              [$cf, 'command_cf_purge_post'] );
+		WP_CLI::add_command( 'servebolt cf purge all',               [$cf, 'command_cf_purge_all'] );
 
-		/*
-		WP_CLI::add_command( 'servebolt cf purge cron status',       [$this, 'command_cf_cron_status'] );
-		WP_CLI::add_command( 'servebolt cf purge cron activate',     [$this, 'command_cf_cron_enable'] );
-		WP_CLI::add_command( 'servebolt cf purge cron deactivate',   [$this, 'command_cf_cron_disable'] );
-		*/
 	}
 
 }
